Added nested tests for setof
These tests required increasing the difference between a View and a View
of a const type. This is because we can't allow a View to simply copy
from a const reference of another View, or we would be able to edit
things we should not be able to.

So we implement a split similar to how iterators and const_iterators are
implemented in the standard library.