Mention explicitly the template parameters of friend Setter.
What we would like is a partial template friendship that matches the
type of both View and Setter, but this is not allowed in C++.

A workaround of making Setter a class nested in another to split its
first template parameter from the second, but would require to use the
class as basically an inned class of the outer shell. This is quite ugly
so we won't do it.

Additionally, this fixes the code not compiling under Visual C++.