Add support for optional SET OF/SEQUENCE OF
This change adds support for optional SET OF/SEQUENCE OF. These are
simply pointers to a normal SET OF/SEQUENCE OF, so we can simply check
if the input name we are getting is a pointer, check if it needs to be
allocated, and continue as normal.

Additionally, two more bugs are fixed.

The first is a naive off-by-one error, where we where checking access to
a list as "< 0 || > size(list)", but it should have been ">=".

The second bug was that setting an item inside a list was not actually
working, due to a namespace missing in a function call.

Simple tests have been added for both cases, and for the new optional
support.